refactor: PostgresDatabase
